feat: add themeId check error

+++ BE/src/main/java/roomescape/reservation/application/ReservationTimeService.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.global.exception.ErrorCode;
 import roomescape.global.exception.ReservationTimeException;
+import roomescape.global.exception.ThemeException;
 import roomescape.reservation.application.dto.ReservationTimeCreateCommand;
 import roomescape.reservation.application.dto.ReservationTimeResult;
 import roomescape.reservation.domain.ReservationRepository;
@@ -47,6 +48,9 @@ public class ReservationTimeService {
         }
         if (reservationRepository.existsByReservationTimeId(id)) {
             throw new ReservationTimeException(ErrorCode.RESERVATION_TIME_ALREADY_USED);
+        }
+        if (reservationRepository.existsByThemeId(id)) {
+            throw new ThemeException(ErrorCode.THEME_ALREADY_USED);
         }
         reservationTimeRepository.deleteById(id);
     }
